commandHauls: extract time stuff to module
